requirements.txt
# Pipecat and AI services
pipecat-ai[daily,silero,elevenlabs,deepgram,tracing]
pipecat-ai-flows
openai
cartesia
daily-python

# Database and storage
pymongo
motor

# Web framework and server
uvicorn
fastapi
requests
httpx
pipecatcloud

# Environment and configuration
python-dotenv

# Testing
pytest
pytest-asyncio

# Audio processing
numpy
scipy

# Observability
langfuse
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http

# Schema and templating
jinja2
pyyaml
pydantic

# Authentication and security
passlib[bcrypt]
python-jose[cryptography]
phonenumbers

# Rate limiting and retry logic
slowapi
tenacity